feat/implementação da primeira tarefa

+++ database/migrations/2023_03_28_172401_create_documents_table.php
@@ -11,18 +11,27 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('documents', function (Blueprint $table) {
-            $table->id();
-            $table->timestamps();
-            $table->bigInteger('category_id');
-            $table->string('title', 60);
-            $table->text('contents');
-
-            $table->foreign('category_id')
-                ->references('id')
-                ->on('categories')
-                ->onDelete('cascade');
-        });
+// A tabela categories já existe na base e por esse motivo ao executar a migration ocorre o erro.
+// Para resolver o problema, basta verificar se a tabela existe antes de criar a mesma.
+// No caso de uma nova instalação, a tabela não existirá e a migration será executada.
+// Caso a tabela já exista, a migration não será executada.
+        if (!Schema::hasTable('categories')) {
+            Schema::create('documents' , function (Blueprint $table) {
+                $table->id();
+                $table->string('title' , 60);
+                $table->text('contents');
+                $table->bigInteger('category_id')->unsigned(); // Adicionei o 'unsigned' para garantir que seja sempre positivo;
+                $table->foreign('category_id')
+                    ->references('id')
+                    ->on('categories')
+                    ->onDelete('cascade');
+                $table->timestamps();
+// Alterei a posição do timestamp para que os campos created_at e updated_at fossem criados no final da tabela.
+// Essa é uma prática comum nos projetos Laravel, tanto nos quais já trabalhei quanto naqueles que acompanho na comunidade.
+// Os campos created_at e updated_at são informações relacionadas à auditoria dos registros e não diretamente relacionadas
+// aos dados principais da tabela.
+            });
+        }
     }
 
     /**
